Adding Coyote support
