+++ src/Test.ConcurrencyTests/Test.ConcurrencyTests.csproj
@@ -59,6 +59,7 @@
 	</ItemGroup>
 
 	<Target Name="CoyoteRewriting" AfterTargets="AfterBuild" Condition=" '$(TargetFramework)'=='netcoreapp3.1' ">
+		<Message Text="This custom target will only be executed when configuration is debug $(TargetFramework)" Importance="high"></Message>
 		<Exec Command="$(ExecCommand)" />
 	</Target>
